code to get sidewalk score and display them on map
This is the updated code to get the sidewalk score and display them on the map the sidewalks are colored so that blue is a good score and red is a bad score with a legend on the map so you can tell what is what. This map is the file called sidewalks_map.html and will be in the data file but you could recreate it here. In addition to this you can hover over a sidewalk to see its SWD_ID and click on it to see its exact score