updated scores file
this updated file fixes the color coding of the sidewalks to accuratley display the scores. In addition it also makes a layer that shows the scores for the neighborhoods in boston. This is created in the neighborhood_score.html file and start with both displays toggled off and you can toggle on the ones that you wish to see